requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
anyio==4.4.0
arrow==1.3.0
asgiref==3.7.2
astroid==3.2.2
binaryornot==0.4.4
black==24.2.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
dill==0.3.8
distlib==0.3.8
dj-database-url==2.1.0
Django==5.0.2
django-cors-headers==4.3.1
django-ninja==1.1.0
django-ratelimit==4.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
ecdsa==0.18.0
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.13.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.6
packaging==24.0
pathspec==0.12.1
pefile==2023.2.7
pendulum==3.0.0
pillow==10.3.0
platformdirs==4.2.0
psycopg2==2.9.9
pyasn1==0.5.1
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
PyJWT==2.8.0
pylint==3.2.2
pylint-django==2.5.5
pylint-plugin-utils==0.8.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pywin32-ctypes==0.2.2
PyYAML==6.0.1
requests==2.32.2
rich==13.7.1
rsa==4.9
setuptools==69.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.37.2
text-unidecode==1.3
time-machine==2.15.0
tomlkit==0.12.5
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.1
uvicorn==0.30.1
virtualenv==20.25.0
watchfiles==0.22.0
websockets==12.0
wheel==0.42.0
